+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 #v0.1.0
 - initial instructions and template for adding questions
+- Questions:
+    - Reverse Integer